Yay! I figured it out using multiple arguments!

+++ exer1.rb
@@ -1,6 +1,9 @@
+# Define a method called double that accepts an argument called my_number and returns that number multiplied by 2.
+# Try calling it multiple times and pass in different numbers each time.
+
 def double(my_num)
     return 2 * my_num
     
 end
 
-p double(16)
+p double(24)
